+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod achievement;
 pub mod achievements;
+pub mod cli;
+pub mod commands;
 pub mod git;
 #[cfg(test)]
 pub mod test;